+++ kitten/mouse picking/MousePicker.h
@@ -6,6 +6,6 @@
 
 namespace MousePicker
 {
-	bool rayHits(const kitten::Ray& p_ray, const kitten::Clickable* p_clickable, float* p_hitOutput);	
+	bool rayHits(const kitten::Ray& p_ray, const kitten::Clickable* p_clickable, double* p_hitOutput);	
 	kitten::Clickable* getClosestHit(const kitten::Ray& p_ray);
 }
